+++ data/saas/dataset/hubspot_dataset.yml
@@ -464,39 +464,47 @@ dataset:
 #            data_categories: [system.operations]
       - name: subscription_preferences
         fields:
-          - name: id
-            data_categories: [system.operations]
-            fidesops_meta:
-              primary_key: True
-              data_type: string
-          - name: name
-            data_categories: [system.operations]
+          - name: recipient
+            data_categories: [user.provided.identifiable.contact.email]
             fidesops_meta:
               data_type: string
-          - name: description
-            data_categories: [system.operations]
+          - name: subscriptionStatuses
             fidesops_meta:
-              data_type: string
-          - name: status
-            data_categories: [system.operations]
-            fidesops_meta:
-              data_type: string
-          - name: sourceOfStatus
-            data_categories: [system.operations]
-            fidesops_meta:
-              data_type: string
-          - name: preferenceGroupName
-            data_categories: [system.operations]
-            fidesops_meta:
-              data_type: string
-          - name: legalBasis
-            data_categories: [system.operations]
-            fidesops_meta:
-              data_type: string
-          - name: legalBasisExplanation
-            data_categories: [system.operations]
-            fidesops_meta:
-              data_type: string
+              data_type: object[]
+            fields:
+              - name: id
+                data_categories: [system.operations]
+                fidesops_meta:
+                  primary_key: True
+                  data_type: string
+              - name: name
+                data_categories: [system.operations]
+                fidesops_meta:
+                  data_type: string
+              - name: description
+                data_categories: [system.operations]
+                fidesops_meta:
+                  data_type: string
+              - name: status
+                data_categories: [system.operations]
+                fidesops_meta:
+                  data_type: string
+              - name: sourceOfStatus
+                data_categories: [system.operations]
+                fidesops_meta:
+                  data_type: string
+              - name: preferenceGroupName
+                data_categories: [system.operations]
+                fidesops_meta:
+                  data_type: string
+              - name: legalBasis
+                data_categories: [system.operations]
+                fidesops_meta:
+                  data_type: string
+              - name: legalBasisExplanation
+                data_categories: [system.operations]
+                fidesops_meta:
+                  data_type: string
 #      - name: users
 #        fields:
 #          - name: id
