+++ docs/fidesops/docs/tutorial/storage_config.md
@@ -34,7 +34,7 @@ def create_local_storage(key, format, access_token):
         json=storage_create_data,
     )
     logger.info(f"Defining an upload location. Status {response.status_code}")
-    return response.json()()
+    return response.json()
 
 ```
